worked on possible inhibit switch/plunger assembly
There're shenanigans to figure out with the arithmetic of where the
switch is placed and the lengths of the plunger/guide. There's also a
slight discrepancy between the reported overtravel, what's shown in the
tech drawing and what's shown in the model (~0.1 mm).

I don't have any more mental energy to put up with SW right now.