started the tail card
The tail card is the position 19 card that serves as a mezzanine for the
-Z endcap. It connects to the -Z solar panels with 12 pin connectors, to
the -Z endcap with a 20 pin and 3 MMCX connectors, and to the backplane
with a 20 pin and 3 MMCX connectors.

While messing with the 12 pin connectors, I noticed that the connectors
for the solar cards are off-center by 0.1 mm (doesn't matter). They're
also a little misplaced in the Z direction. We'll have to go in and
tweak their Z position later.