added cutouts; reduced number of screw pillars
The cutouts are for reducing mass. We could get away with removing a
little more mass from the spaces with no screw pillars. However, it
would make things a little less uniform. (annoying if we want to be able
to manually machine this)

Each side in the X axis now only uses 4 screws to fasten together the
rack. Additionally, these screw holes do not intersect with the holes
that fasten the solar panels, so everything can be through-holes!