symmetrized solar holes; decoupled struts & slots
The struts are now a feature on a separate "level" from the hold-down
slots, so they don't need to be placed in any specific location relative
to the slots. (In fact, they can be right on top of the slots, which is
nice for future changes.)

The holes that mount the solar panel to the frame are now symmetrical
along the Z-axis.