Tweaked positions of the spring pins and lip
The spring pins now have an adequate margin away from the thermal contact area,
so that they won't bleed solder into it. The thermal contact area has been expanded
to account for the tolerance in the card slot. These changes brought the pins very
close to the edge of the lip. So, the lip was extended by 0.3 mm and the version
number on the side panels was bumped to 1.0.1.