created a debug breakout board
The +Z endcard now has a connector to allow direct connection to the CAN
bus from outside the satellite  while it's fully assembled. There is a
corresponding cutout in the +Z endcap. The debug board gets fastened to
the structure using a threaded hole that would normally be used for the
+Z endcap. This requires a 2 mm spacer.