moved screws together to prevent warping
The batteries were being compressed unevenly [we think]. The outside
screws were close to the edge of the battereis, so they had less
battery to compress. This made the batteries lemon-shaped and caused the
board to go out of flat. Bringing the screws closer together evens out
the length of battery that each screw needs to compress, so it should
reduce the warping of the batteries/board.