Fixed Assembly, added fillet to ramps, updated version #
After I reached out to a local machine shop they told me to add some fillets where the ramps meet the base of the frame. I updated the verison numbers to 2.0.0, and did todays date becuase hopefully they get sent out to be manufactured today. Also I added a fillet to the wedges, since with the new fillet on the ramp, they would collide without it. The assembly isnt fully complete, I still need to add some fasteners, but as of now there are no new conflicts due to the design change, there are old conflicts, like the bottom star tracker and the battery, and the reaction wheels with eachother, but nothing new.